Support show_meeting_attendees in v1_meeting
This is a new field being used by LFX One to show data about
meeting attendees to other meeting attendees. We need to display this
even for v1 meetings, so add to the sync helper to tack it on when it is
a board meeting and the project is tlf. Otherwise set to false for now.

Signed-off-by: Jordan Evans <jevans@linuxfoundation.org>